#AVO-3636: Move content of quicklane modal and share through email modal to separate components. Create new modal to use both of these new components.
