+++ src/item/views/ItemDetail.tsx
@@ -57,7 +57,6 @@ import {
 	type LoadingInfo,
 	ShareThroughEmailModal,
 } from '../../shared/components';
-import QuickLaneModal from '../../shared/components/QuickLaneModal/QuickLaneModal';
 import { LANGUAGES, ROUTE_PARTS } from '../../shared/constants';
 import {
 	buildLink,
@@ -110,6 +109,7 @@ import ReportItemModal from '../components/modals/ReportItemModal';
 import { RELATED_ITEMS_AMOUNT } from '../item.const';
 import { type ItemTrimInfo } from '../item.types';
 import './ItemDetail.scss';
+import FragmentShareModal from '../../shared/components/FragmentShareModal/FragmentShareModal';
 
 interface ItemDetailProps {
 	id?: string; // Item id when component needs to be used inside another component and the id cannot come from the url (match.params.id)
@@ -171,9 +171,8 @@ const ItemDetail: FC<ItemDetailProps & DefaultSecureRouteProps<{ id: string }>> 
 	const [loadingInfo, setLoadingInfo] = useState<LoadingInfo>({ state: 'loading' });
 	const [isAddToCollectionModalOpen, setIsAddToCollectionModalOpen] = useState(false);
 	const [isAddToFragmentModalOpen, setIsAddToFragmentModalOpen] = useState(false);
-	const [isShareThroughEmailModalOpen, setIsShareThroughEmailModalOpen] = useState(false);
 	const [isReportItemModalOpen, setIsReportItemModalOpen] = useState(false);
-	const [isQuickLaneModalOpen, setIsQuickLaneModalOpen] = useState(false);
+	const [isShareFragmentModalOpen, setIsShareFragmentModalOpen] = useState(false);
 	const [relatedItems, setRelatedItems] = useState<Avo.Search.ResultItem[] | null>(null);
 	const [bookmarkViewPlayCounts, setBookmarkViewPlayCounts] = useState<BookmarkViewPlayCounts>(
 		DEFAULT_BOOKMARK_VIEW_PLAY_COUNTS
@@ -792,21 +791,16 @@ const ItemDetail: FC<ItemDetailProps & DefaultSecureRouteProps<{ id: string }>> 
 							</DropdownContent>
 						</Dropdown>
 					)}
-
-					{PermissionService.hasPerm(commonUser, PermissionName.CREATE_QUICK_LANE) && (
-						<Button
-							type="tertiary"
-							icon={IconName.link2}
-							label={tText('item/views/item___delen-met-leerlingen')}
-							ariaLabel={tText(
-								'item/views/item-detail___deel-dit-met-alle-leerlingen'
-							)}
-							title={tText('item/views/item-detail___deel-dit-met-alle-leerlingen')}
-							onClick={() => {
-								setIsQuickLaneModalOpen(true);
-							}}
-						/>
-					)}
+					<Button
+						type="tertiary"
+						icon={IconName.share2}
+						label={tText('Deel dit fragment')}
+						ariaLabel={tText('Deel dit fragment')}
+						title={tText('Deel dit fragment')}
+						onClick={() => {
+							setIsShareFragmentModalOpen(true);
+						}}
+					/>
 					{PermissionService.hasPerm(commonUser, PermissionName.VIEW_ITEMS_OVERVIEW) && (
 						<Link to={buildLink(ITEMS_PATH.ITEM_DETAIL, { id: item?.uid })}>
 							<Button
@@ -830,17 +824,6 @@ const ItemDetail: FC<ItemDetailProps & DefaultSecureRouteProps<{ id: string }>> 
 							title: tText('item/views/item___toggle-bladwijzer'),
 							onClick: toggleBookmark,
 						})}
-
-					<Button
-						type="tertiary"
-						icon={IconName.share2}
-						ariaLabel={tText('item/views/item___share-item')}
-						title={tText('item/views/item___share-item')}
-						onClick={() => {
-							setIsShareThroughEmailModalOpen(true);
-						}}
-					/>
-
 					<Button
 						type="tertiary"
 						icon={IconName.flag}
@@ -940,11 +923,7 @@ const ItemDetail: FC<ItemDetailProps & DefaultSecureRouteProps<{ id: string }>> 
 							itemMetaData={item}
 							showMetadata={false}
 							enableMetadataLink={false}
-							canPlay={
-								!isAddToCollectionModalOpen &&
-								!isShareThroughEmailModalOpen &&
-								!isReportItemModalOpen
-							}
+							canPlay={!isAddToCollectionModalOpen && !isReportItemModalOpen}
 							verticalLayout={isMobileWidth()}
 							cuePointsVideo={{
 								start: cuePoint ? parseInt(cuePoint.split(',')[0], 10) : null,
@@ -1020,18 +999,6 @@ const ItemDetail: FC<ItemDetailProps & DefaultSecureRouteProps<{ id: string }>> 
 						/>
 					)}
 				{!renderActionButtons && (
-					<ShareThroughEmailModal
-						modalTitle={tText('item/views/item-detail___deel-dit-item')}
-						type="item"
-						emailLinkHref={window.location.href}
-						emailLinkTitle={item.title}
-						isOpen={isShareThroughEmailModalOpen}
-						onClose={() => {
-							setIsShareThroughEmailModalOpen(false);
-						}}
-					/>
-				)}
-				{!renderActionButtons && (
 					<ReportItemModal
 						externalId={match.params.id}
 						isOpen={isReportItemModalOpen}
@@ -1040,17 +1007,17 @@ const ItemDetail: FC<ItemDetailProps & DefaultSecureRouteProps<{ id: string }>> 
 						}}
 					/>
 				)}
-				{PermissionService.hasPerm(commonUser, PermissionName.CREATE_QUICK_LANE) && (
-					<QuickLaneModal
-						modalTitle={tHtml('item/views/item___snel-delen-met-leerlingen')}
-						isOpen={isQuickLaneModalOpen}
-						content={item}
-						content_label="ITEM"
-						onClose={() => {
-							setIsQuickLaneModalOpen(false);
-						}}
-					/>
-				)}
+				<FragmentShareModal
+					isOpen={isShareFragmentModalOpen}
+					item={item}
+					withPupils={PermissionService.hasPerm(
+						commonUser,
+						PermissionName.CREATE_QUICK_LANE
+					)}
+					onClose={() => {
+						setIsShareFragmentModalOpen(false);
+					}}
+				/>
 				{PermissionService.hasPerm(commonUser, PermissionName.CREATE_ASSIGNMENTS) && (
 					<ImportToAssignmentModal
 						isOpen={isImportToAssignmentModalOpen}
