+++ src/shared/components/FragmentShareModal/FragmentShareModal.scss
@@ -1,0 +1,17 @@
+@import "../../../styles/settings/colors";
+@import "../../../styles/settings/variables";
+
+.m-fragment-share-modal__tabs-wrapper {
+	// Equate to avo2-components/src/components/Modal/Modal.scss without importing
+	margin: 0 - ($g-spacer-unit * 3);
+	padding: 0 ($g-spacer-unit * 1.5); // ensure text is aligned, 12 + 12
+	background-color: $color-gray-50;
+
+	.c-tab-item.c-tab-item__active {
+		border-bottom:  0.2rem solid $color-gray-200;
+		padding: 1.5rem 0;
+		margin-right: 1.7rem;
+		margin-left: 1.7rem;
+		margin-bottom: 0;
+	}
+}
