+++ src/shared/components/QuickLaneContent/QuickLaneContent.helpers.tsx
@@ -3,7 +3,7 @@ import { type Avo } from '@viaa/avo2-types';
 import { AssignmentLayout } from '../../../assignment/assignment.types';
 import { type QuickLaneUrlObject } from '../../types';
 
-import { type QuickLaneType } from './QuickLaneModal.types';
+import { type QuickLaneType } from './QuickLaneContent.types';
 
 export const isShareable = (
 	content: Avo.Assignment.Assignment | Avo.Collection.Collection | Avo.Item.Item
