+++ src/shared/components/QuickLaneContent/QuickLaneContent.scss
@@ -4,7 +4,7 @@
 	width: $m-content-thumbnail-width;
 }
 
-.m-quick-lane-modal {
+.m-quick-lane-content {
 	&__avatar {
 		.c-avatar--img {
 			border: 1px solid $color-night-blue;
