+++ src/shared/components/ShareThroughEmailContent/ShareThroughEmailContent.scss
